* subversion/libsvn_subr/gpg_agent.c   (receive_from_gpg_agent): Eliminate integer size mismatch warning.
git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@1396402 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/libsvn_subr/gpg_agent.c
@@ -107,7 +107,7 @@ static svn_boolean_t
 receive_from_gpg_agent(int sd, char *buf, size_t n)
 {
   int i = 0;
-  int recvd;
+  size_t recvd;
   char c;
 
   /* Clear existing buffer content before reading response. */
